pyproject.toml
[project]
name = "client-onboarding-rag-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faker>=37.6.0",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-openai>=0.3.32",
    "python-dotenv>=1.1.1",
    "streamlit>=1.49.1",
]
